データインポート処理の改善と型定義の追加、可視化用の色とレイアウト設定を新規作成しました。また、不要なコードを削除し、クライアントコンポーネントのインポートを整理しました。
