+++ src/theme/rateColor.ts
@@ -1,0 +1,24 @@
+const colors = [
+    '#1f77b4', // 青
+    '#ff7f0e', // オレンジ
+    '#2ca02c', // 緑
+    '#d62728', // 赤
+    '#9467bd', // 紫
+    '#8c564b', // 茶色
+    '#e377c2', // ピンク
+    '#7f7f7f', // グレー
+    '#bcbd22', // 黄緑
+    '#17becf', // 水色
+    '#aec7e8', // 薄い青
+    '#ffbb78', // 薄いオレンジ
+    '#98df8a', // 薄い緑
+    '#ff9896', // 薄い赤
+    '#c5b0d5', // 薄い紫
+    '#c49c94', // 薄い茶色
+    '#f7b6d2', // 薄いピンク
+    '#c7c7c7', // 薄いグレー
+    '#dbdb8d', // 薄い黄緑
+    '#9edae5'  // 薄い水色
+];
+
+export default colors;
